+++ xpath-to-xml-dom/src/main/java/com/github/simy4/xpath/dom/spi/DomNavigatorSpi.java
@@ -5,6 +5,7 @@ import com.github.simy4.xpath.dom.navigator.DomNavigator;
 import com.github.simy4.xpath.dom.navigator.DomNode;
 import com.github.simy4.xpath.spi.Effect;
 import com.github.simy4.xpath.spi.NavigatorSpi;
+import org.w3c.dom.Document;
 import org.w3c.dom.Node;
 
 /**
@@ -14,7 +15,7 @@ public class DomNavigatorSpi implements NavigatorSpi {
 
     @Override
     public boolean canHandle(Object o) {
-        return o instanceof Node;
+        return o instanceof Node && (o instanceof Document || (null != ((Node) o).getOwnerDocument()));
     }
 
     @Override