+++ xpath-to-xml-jdom/src/main/java/module-info.java
@@ -1,0 +1,6 @@
+module com.github.simych.xpath.jdom {
+    requires transitive com.github.simych.xpath.core;
+    requires jdom2;
+
+    provides com.github.simy4.xpath.spi.NavigatorSpi with com.github.simy4.xpath.jdom.spi.JDomNavigatorSpi;
+}